Move page-templates-template-parts to page-templates
