+++ packages/edit-site/src/components/page-templates/style.scss
