+++ packages/edit-site/src/components/sidebar-navigation-screen-templates-browse/content.js
@@ -9,7 +9,7 @@ import { __experimentalItemGroup as ItemGroup } from '@wordpress/components';
  * Internal dependencies
  */
 import DataViewItem from '../sidebar-dataviews/dataview-item';
-import { useAddedBy } from '../page-templates-template-parts/hooks';
+import { useAddedBy } from '../page-templates/hooks';
 import { layout } from '@wordpress/icons';
 
 const EMPTY_ARRAY = [];
